Migrate core/permissions to ts
