+++ packages/core/permissions/.eslintignore
@@ -1,2 +1,4 @@
 node_modules/
 .eslintrc.js
+dist/
+jest.config.js
