+++ packages/core/permissions/jest.config.js
@@ -2,5 +2,6 @@
 
 module.exports = {
   preset: '../../../jest-preset.unit.js',
+  testMatch: ['**/__tests__/**/*.test.ts'],
   displayName: 'Core permissions',
 };
