--- packages/core/permissions/lib/domain/index.js
@@ -1,7 +1,0 @@
-'use strict';
-
-const permission = require('./permission');
-
-module.exports = {
-  permission,
-};
