--- packages/core/permissions/lib/engine/abilities/index.js
@@ -1,7 +1,0 @@
-'use strict';
-
-const { caslAbilityBuilder } = require('./casl-ability');
-
-module.exports = {
-  caslAbilityBuilder,
-};
