--- packages/core/permissions/lib/index.js
@@ -1,9 +1,0 @@
-'use strict';
-
-const domain = require('./domain');
-const engine = require('./engine');
-
-module.exports = {
-  domain,
-  engine,
-};
