+++ packages/core/permissions/package.json
@@ -19,8 +19,17 @@
       "url": "https://strapi.io"
     }
   ],
-  "main": "./lib/index.js",
+  "files": [
+    "./dist"
+  ],
+  "main": "./dist/index.js",
+  "types": "./dist/index.d.ts",
   "scripts": {
+    "build": "run -T tsc",
+    "build:ts": "run build",
+    "watch": "run -T tsc -w --preserveWatchOutput",
+    "clean": "run -T rimraf ./dist",
+    "prepublishOnly": "yarn clean && yarn build",
     "test:unit": "run -T jest",
     "test:unit:watch": "run -T jest --watch",
     "lint": "run -T eslint ."
@@ -30,6 +39,10 @@
     "@strapi/utils": "4.11.7",
     "lodash": "4.17.21",
     "sift": "16.0.1"
+  },
+  "devDependencies": {
+    "eslint-config-custom": "4.11.7",
+    "tsconfig": "4.11.7"
   },
   "engines": {
     "node": ">=14.19.1 <=18.x.x",
