+++ packages/core/permissions/src/domain/index.ts
@@ -1,0 +1,3 @@
+import * as permission from './permission';
+
+export { permission };
