+++ packages/core/permissions/src/engine/abilities/index.ts
@@ -1,0 +1,1 @@
+export * from './casl-ability';
