+++ packages/core/permissions/src/index.ts
@@ -1,0 +1,7 @@
+import * as domain from './domain';
+import * as engine from './engine';
+
+export = {
+  domain,
+  engine,
+};
