+++ packages/core/permissions/tsconfig.eslint.json
@@ -1,0 +1,8 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "noEmit": true
+  },
+  "include": ["types", "src"],
+  "exclude": ["node_modules"]
+}
