+++ packages/core/permissions/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "tsconfig/base.json",
+  "compilerOptions": {
+    "outDir": "dist"
+  },
+  "include": ["types", "src"],
+  "exclude": ["node_modules", "**/__tests__/**"]
+}
